Change SourceModel::readXml() to eliminate cloning of "Sources"s when they are read in. This is primarily to optimize memory usage with large diffuse models.
SourceModel::addSource() changed to accept a "fromClone" option which
directs it to clone the source being passed to it (the default for
compatibility with how it worked before) or to adopt the Source being
passed (i.e. to take over its memory management). This change is
propagated to the derived classes LogLike and BinnedLikelihood.

A new function SourceFactory::releaseSource also added to return the
private Source instance from SourceFactory without cloning. Using this
function make SourceFactory "forget" about that source and the user takes
over responsibility for ultimate deletion.